[utils] Fold warn tests into the createLogOnce tests
